requirements.txt

langchain>=0.2.14
langchain-openai>=0.1.7
python-dotenv>=1.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
